Skip fused-rmsnorm in QKV kernel if rmsnorm object is None
The QKV kernel's fused_rmsnorm parameter is set to true when not doing
sequence_parallel.  This is correct for all llama models except for Eagle draft
which does not have an RMSNorm before the QKV projection.

Instead of passing in an `is_eagle_draft` flag all the way in from model to
attention_base to gqa, we can check whether the `rmsnorm` object is None to
know if fused-rmsnorm should be performed.  The rmsnorm object is created in
llama model class and set to None for eagle_draft.

cr: https://code.amazon.com/reviews/CR-158277179